requirements.txt
python-binance
django-searchable-encrypted-fields
